own css, ruleset list as divs, spinners
I just took the css from andymatthews page because it was having issues
and not loading on the page. I also recognized our lists were all in
divs except for the ruleset list, so I made it so that was consistent
and also made it a collapsible list so we could make it more efficient
later on. I also made a function for the spinners to differ on whether
it’s loading or updating.