delete and flush buttons in manage-list
Removed the delete and flush from the update-url page. Added them to
the collapsable set instead. One bug with delete button having to go to
home page at the moment. Issue when trying to just have it refresh the
page and it goes to the about empty page. Reoccurring problem needing
to be taken care of.