alert for issue #58
I created a function that calls devtools.showInstalledRulesets and
checks for a installed ruleset that shares ID. the user is alerted if
found. this will slow things down because of the extra call, we could
put a value in window to store this from the window before, or access
the displayed information.