Prototype Management in Devtools
Adding prototype manager page to Devtools site. Provides a listing of
all prototypes and the ability to edit and delete them. The page will
also allow adding new prototypes.

This commit also adds the selection of a prototype when creating a
child pico in the #about page.

Adding, editing, and deleting prototypes are currently non-functional.